Support lists of request, Support custom output json
Each usecase defines the name of the output variable.
This variable can be modified by the user request, by
mapping it to another custom variable name

The current usecases support a single output variable, since they are
using the pass through service.
In next iterations, we can have more services for text completion,
which might result in more output variables.
The type of the service, should then be added to the configuration.