Allow for flattening the text completion response params
Instead of a list of PromptParams of Dict[str,str],
we allow to return a list of one PromptParams of Dict[str, List[str]].

This will allow us to chain usecases with a map reduce approach.

For example, we can run single-doc summaries usecase of many docs,
and then chain it with a multi-doc summary usecase which takes
all the single-doc summaries as input.