Handle all possible combinations of usecases in one flow using recursion
We now have one usecase structure which can define a list of usecases to
be executed as a chain or in parallel.
Each nested usecase can be another chain or parallel list of other
lower-level usecases.

There is one executor per case: single usecase, chain, parallel.
All executors return the same type back in order to be able to
combine them as a series of steps that modify the list of output params
along the way.