Introduce TextCompletionServiceFactory and declare service_name in each usecase config
Now the usecase config is able to declare the service_name it should use
to perform the text completion.

Currently pass_through service is supported which calls the text completion client
and passes through the response without postprocessing.

This change was made so we can support different services which
handle the output of the text completion clients in a different way.